clean up

+++ .evergreen/orchestration/setup.sh
@@ -2,17 +2,24 @@
 # Handle setup for orchestration
 
 set -o errexit
-
 SCRIPT_DIR="$(dirname "${BASH_SOURCE[0]}")"
 . "${SCRIPT_DIR:?}/../handle-paths.sh"
 
-(
-  cd $SCRIPT_DIR/..
-  bash init-node-and-npm-env.sh
-)
+if [ -n "${CI:-}" ]; then
+  (
+    cd $SCRIPT_DIR/..
+    bash ../install-node.sh
+    source ../init-node-and-npm-env.sh
 
-(
-  cd $SCRIPT_DIR
-  npm install .
-  npm run compile
-)
+    # Use the standard registry.
+    npm config set -L project registry "https://registry.npmjs.org"
+  )
+fi
+
+if [ ! -f $SCRIPT_DIR/index.js ]; then
+  (
+    cd $SCRIPT_DIR
+    npm install .
+    npm run compile
+  )
+fi
